Fixes to Coverity defects 85010 84878 72752 72742 72719 and 73418. skygw_utils.cc: Added function is_valid_posix_path that checks if a path is POSIX-compliant.

+++ server/core/test/testadminusers.c
@@ -272,6 +272,8 @@ char	*home, buf[1024];
 	if ((home = getenv("MAXSCALE_HOME")) == NULL || strlen(home) >= 1024)
 		home =  "/usr/local/skysql";
 	sprintf(buf, "%s/etc/passwd", home);
+    if(!is_valid_posix_path(buf))
+        exit(1);
 	if (strcmp(buf, "/etc/passwd") != 0)
 		unlink(buf);
